move service implementation to addon namespace
